backend/requirements.txt
python-dotenv==1.0.0
asyncpg==0.29.0
aiofiles==23.2.0
fastapi==0.104.1
uvicorn[standard]==0.24.0